requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2020.6.20
Django==4.0.3
django-heroku==0.3.1
django-cors-headers==3.11.0
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
mysqlclient==2.1.0
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
whitenoise==6.0.0
yapf==0.32.0
